feature[test]: refactoring issue #109
